+++ magnus-macros/src/init.rs
@@ -27,13 +27,19 @@ pub fn expand(args: AttributeArgs, input: ItemFn) -> TokenStream {
     );
     let init_name = input.sig.ident.clone();
 
+    let init_type = if input.sig.inputs.is_empty() {
+        quote!(magnus::method::Init)
+    } else {
+        quote!(magnus::method::InitRuby)
+    };
+
     quote! {
         #input
 
         #[allow(non_snake_case)]
         #[no_mangle]
         pub extern "C" fn #extern_init_name() {
-            unsafe { magnus::method::Init::new(#init_name).call_handle_error() }
+            unsafe { #init_type::new(#init_name).call_handle_error() }
         }
     }
 }