+++ CHANGELOG.md
@@ -94,6 +94,8 @@
 - The `wb_protected` flag is automatically set for data wrapped in a Ruby
   object when `mark` is not set.
 - `TypedData::class` and `TypedData::class_for` now take a `&Ruby` argument.
+- init function marked with `#[magnus::init]` optionally takes a `&Ruby`
+  argument.
 
 ### Deprecated
 - `RString::append` (use `RString::buf_append`).